[WIP] Chromium: de-linuxify
The macOS situation is far too complicated for now. Parking
this. libsystem 10.11.6 declares the macos version to be 10.11,
chromium uses this to check which features to backport/polyfill. In
particular, the NSPressureConfiguration[1] isn't backported when
building on 10.11. This trivially fails, because the headers are
actually 10.10. The backport is disabled, but the definition is
absent.

[1] which Apple claims is part of 10.10.3+, but I have yet to be able
to verify that the headers for 10.10 support this. Maybe the headers
were introduced in 10.11, but the feature was silently present in
10.10.3? Wild conjecture and archaeology ensues.