[WIP] bootstrap a compiler-rt without libcxx
Why do this instead of building a compiler-rt against the bootstrap
versions of libcxx? I guess it's more pure, and probably makes more
sense when bootstrap clang and runtime clang aren't the same version

So the dependencies are:

  compiler-rt-bootstrap -> libcxxabi -> libcxx -> compiler-rt (full)

TODO: check this doesn't cause two instances of compiler-rt. Is it
even a dynamic library?