Use the compiler-rt from bootstrap tools in place of bootstrapping a compiler-rt
Not clear if there's any purity to be won by rebuilding compiler-rt
twice. In fact, in may reduce correctness. We assume that compiler-rt
and clang versions should be matched, so wherever we use bootstrap
clang we should also use bootstrap compiler-rt. This leaves stage 4 in
an awkward spot because we switch over to rebuilt compiler-rt without
switching over clang.