[WIP] park my work state
There's something clever to be done to bootstrap the libc++abi,
libc++, and clang chain. This way _might_ work, but after implementing
this with two extra stages (!!) I'm inclined to believe that the
llvmPackages inspired single-stage bootstrap is much easier to
understand, less verbose, and more correct.

In particular, when stdenv.hostPlatform.useLLVM is set, then libc++
and libc++abi are built with lldClangNoLibcxx, and compiler-rt is
build from lldClangNoCompilerRt. This seems to be a cycle-breaking
bootstrap that uses a working cross compiler to produce the required
libraries. For the bootstrap build we are in a similar situation.

Note that the general case of building llvmPackages_m from
llvmPackages_n seems to cause references to libc++ to be retained from
llvmPackages_n.libcxx, which isn't at all surprising, but not
acceptable when trying to discharge bootstrap dependencies.

While we're here, also pay attention to the clang in bootstrap tools
being inserted into llvmPackages_7, and assembled by the same wrapper
as the rebuilt clang. This is a lie since bootstrap clang is
v4.0.1. Furthermore, this is a problem when trying to define an
equivalent to lldClangNoCompilerRt, since current bootstrap clang
4.0.1 doesn't support passing `-nostdlib++`.

TL;DR: I think what I had at the start of this branch was good, though
the branch also includes various improvements it went off the rails in
the last few commits.