[wip] make a bit of a mess of CF
This is mostly working around not being able to build
swift-corefoundation.

It's tainted by *something* leaking bootstrap-tools itself as a
dependency into stage1, so ${bootstrap-}tools/Library/Frameworks (and
probably ${bootstrap-tools}/lib and ${bootstrap-tools}/include) are
present until this is gone. It has to be one of binutils, coreutils,
or gnugrep.